Correct clog10 documentation (bug 19673).
Bug 19673 reports that the documentation of clog10 is incorrect, both
failing to include the division by log (10) in the imaginary part and,
in the non-TeX version of the equation only, describing the LHS as log
rather than log10.

This patch fixes both issues.  Note: I think it's appropriate that the
LHS says log10 not clog10, and that the cexp and clog descriptions
referred to in a comment in that bug report similarly say exp and log;
this is a mathematical description not a literal C one.

Tested for x86_64.

	[BZ #19673]
	* manual/math.texi (Exponents and Logarithms): Correct description
	of clog10.