Fix SH4 FP_ILOGB0 (bug 20859).
ISO C requires that the value of FP_ILOGB0 must be INT_MIN or
-INT_MAX.  In sysdeps/sh/sh4/bits/mathdef.h, it's 0x80000001; that is,
a positive unsigned value that would be -INT_MAX if converted to int,
which is not valid (there's no actual constraint on the type, but
whatever the type the integer value must be one of the two permitted,
and types other than int don't really make sense).  This patch makes
the ABI-compatible change to (-0x7fffffff).

(The testcase handles positive and negative sign separately to avoid
any issues with implicit conversions that could result in e.g. INT_MIN
converted to uintmax_t wrongly passing.)

Tested (compilation only) with build-many-glibcs.py.

	[BZ #20859]
	* sysdeps/sh/sh4/bits/mathdef.h (FP_ILOGB0): Define to
	(-0x7fffffff) instead of 0x80000001.
	* math/test-fp-ilogb-constants.c: New file.
	* math/Makefile (tests): Add test-fp-ilogb-constants.