Use Linux 4.9 (headers) in build-many-glibcs.py.
This patch updates build-many-glibcs.py to use Linux 4.9 for kernel
headers unless another version is explicitly specified.  Note that
when a version changes like this you'll need to use --replace-sources
when updating an existing checkout to tell build-many-glibcs.py it's
OK to delete and replace the sources of a component for which the
version used has changed.

	* scripts/build-many-glibcs.py (Context.checkout): Default Linux
	kernel version to 4.9.