Fix generic fmax, fmin sNaN handling (bug 20947).
Various fmax and fmin function implementations mishandle sNaN
arguments:

(a) When both arguments are NaNs, the return value should be a qNaN,
but sometimes it is an sNaN if at least one argument is an sNaN.

(b) Under TS 18661-1 semantics, if either argument is an sNaN then the
result should be a qNaN (whereas if one argument is a qNaN and the
other is not a NaN, the result should be the non-NaN argument).
Various implementations treat sNaNs like qNaNs here.

This patch fixes the generic implementations used in the absence of
architecture-specific versions.

Tested for mips64 and powerpc (together with testcases that I'll add
along with the x86_64 / x86 fixes).

	[BZ #20947]
	* math/s_fmax_template.c (M_DECL_FUNC (__fmax)): Add the arguments
	when either is a signaling NaN.
	* math/s_fmin_template.c (M_DECL_FUNC (__fmin)): Likewise.